+++ app/src/main/java/ru/practicum/android/diploma/common/di/DataModule.kt
@@ -17,7 +17,7 @@ import ru.practicum.android.diploma.common.data.network.RetrofitNetworkClient
 import ru.practicum.android.diploma.common.data.storage.FilterStorage
 import ru.practicum.android.diploma.common.data.storage.SharedPrefsFilterStorage
 import ru.practicum.android.diploma.common.data.storage.SharedPrefsFilterStorage.Companion.SEARCH_FILTER_PREFERENCES
-import ru.practicum.android.diploma.common.data.storage.mapper.FilterMapper
+import ru.practicum.android.diploma.common.mappers.FilterMapper
 import ru.practicum.android.diploma.search.data.mapper.VacancyResponseMapper
 
 val dataModule = module {
