+++ app/src/main/java/ru/practicum/android/diploma/filter/data/dto/AreaDto.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class AreaDto(
     val id: String,
+    val name: String,
     @SerializedName("parent_id")
     val parentId: String,
-    val name: String,
 )
