+++ app/src/main/java/ru/practicum/android/diploma/filter/data/dto/AreaResponse.kt
@@ -3,5 +3,5 @@ package ru.practicum.android.diploma.filter.data.dto
 import ru.practicum.android.diploma.common.data.network.dto.Response
 
 data class AreaResponse(
-    val areas: List<Areas>
+    val areas: List<AreasDto>
 ) : Response()
