+++ app/src/main/java/ru/practicum/android/diploma/filter/data/dto/AreasDto.kt
@@ -1,5 +1,5 @@
 package ru.practicum.android.diploma.filter.data.dto
 
-data class Areas(
+data class AreasDto(
     val areas: List<AreaDto>
 )
